+++ src/s2geography/geoarrow-geography.h
@@ -7,7 +7,7 @@
 
 #include <vector>
 
-#include "geoarrow/geoarrow.hpp"
+#include "s2geography/geoarrow-geography_util.h"
 
 namespace s2geography {
 
@@ -41,13 +41,13 @@ class GeoArrowPointShape : public S2Shape {
   /// \brief Create an empty point shape representing zero points
   GeoArrowPointShape() = default;
 
-  struct GeoArrowGeometryView nodes() const { return geom_; }
-
   /// \brief Convenience constructor that initializes a point shape
   ///
   /// Throws if geom neither a POINT nor a MULTIPOINT, or is a MULTIPOINT
   /// that contains EMPTY children.
   explicit GeoArrowPointShape(struct GeoArrowGeometryView geom);
+
+  GeoArrowGeom geom() const { return geom_; }
 
   /// \brief Reset internal state such that this shape represents zero edges
   void Clear();
@@ -72,7 +72,7 @@ class GeoArrowPointShape : public S2Shape {
   TypeTag type_tag() const override;
 
  private:
-  struct GeoArrowGeometryView geom_{};
+  GeoArrowGeom geom_;
 };
 
 /// \brief Linestring S2Shape implementation backed by a GeoArrowGeometryView
@@ -94,7 +94,7 @@ class GeoArrowLaxPolylineShape : public S2Shape {
   /// Throws if geom neither a LINESTRING nor a MULTILINESTRING.
   explicit GeoArrowLaxPolylineShape(struct GeoArrowGeometryView geom);
 
-  struct GeoArrowGeometryView nodes() const { return geom_; }
+  GeoArrowGeom geom() const { return geom_; }
 
   /// \brief Reset internal state such that this shape represents zero edges
   void Clear();
@@ -115,7 +115,7 @@ class GeoArrowLaxPolylineShape : public S2Shape {
   TypeTag type_tag() const override;
 
  private:
-  struct GeoArrowGeometryView geom_{};
+  GeoArrowGeom geom_{};
   int num_chains_{};
   std::vector<int> num_edges_;
 };
@@ -151,7 +151,7 @@ class GeoArrowLaxPolygonShape : public S2Shape {
   /// Throws if geometry is not a POLYGON or MULTIPOLYGON.
   explicit GeoArrowLaxPolygonShape(struct GeoArrowGeometryView geom);
 
-  struct GeoArrowGeometryView nodes() const { return geom_; }
+  GeoArrowGeom geom() const { return geom_; }
 
   /// \brief Reset internal state such that this shape represents zero edges
   void Clear();
@@ -184,7 +184,7 @@ class GeoArrowLaxPolygonShape : public S2Shape {
   TypeTag type_tag() const override;
 
  private:
-  struct GeoArrowGeometryView geom_{};
+  GeoArrowGeom geom_{};
   int num_loops_{};
   // Cumulative edge counts per loop: num_edges_[0] = 0,
   // num_edges_[i+1] = total edges in loops 0..i
@@ -286,151 +286,6 @@ class GeoArrowGeography {
 
   void InitIndex();
   void GetCellUnionBound(std::vector<S2CellId>* cell_ids);
-};
-
-namespace internal {
-
-template <typename Visit>
-void VisitGeoArrowNodes(struct GeoArrowGeometryView geom, Visit&& visit) {
-  if (geom.size_nodes == 0) {
-    return;
-  }
-
-  const struct GeoArrowGeometryNode* end = geom.root + geom.size_nodes;
-  for (const struct GeoArrowGeometryNode* node = geom.root; node < end;
-       ++node) {
-    visit(node);
-  }
-}
-
-template <typename Visit>
-void VisitLngLat(const struct GeoArrowGeometryNode* node, int64_t offset,
-                 int64_t n, Visit&& visit) {
-  const uint8_t* lngs = node->coords[0] + offset * node->coord_stride[0];
-  const uint8_t* lats = node->coords[1] + offset * node->coord_stride[1];
-  double lng, lat;
-
-  if (node->flags & GEOARROW_GEOMETRY_NODE_FLAG_SWAP_ENDIAN) {
-    uint64_t tmp;
-    for (int64_t i = 0; i < n; ++i) {
-      memcpy(&tmp, lngs, sizeof(double));
-      tmp = GEOARROW_BSWAP64(tmp);
-      memcpy(&lng, &tmp, sizeof(double));
-
-      memcpy(&tmp, lats, sizeof(double));
-      tmp = GEOARROW_BSWAP64(tmp);
-      memcpy(&lat, &tmp, sizeof(double));
-      visit(lng, lat);
-
-      lngs += node->coord_stride[0];
-      lats += node->coord_stride[1];
-    }
-  } else {
-    for (int64_t i = 0; i < n; ++i) {
-      memcpy(&lng, lngs, sizeof(double));
-      memcpy(&lat, lats, sizeof(double));
-      visit(lng, lat);
-
-      lngs += node->coord_stride[0];
-      lats += node->coord_stride[1];
-    }
-  }
-}
-
-template <typename Visit>
-void VisitVertices(const struct GeoArrowGeometryNode* node, Visit&& visit) {
-  if (node->size < 2) {
-    return;
-  }
-
-  VisitLngLat(node, 0, node->size, [&](double lng0, double lat0) {
-    visit(S2LatLng::FromDegrees(lat0, lng0).ToPoint());
-  });
-}
-
-template <typename Visit>
-void VisitEdges(const struct GeoArrowGeometryNode* node, Visit&& visit) {
-  if (node->size < 2) {
-    return;
-  }
-
-  S2Shape::Edge e;
-  VisitLngLat(node, 0, node->size - 1, [&](double lng0, double lat0) {
-    VisitLngLat(node, 1, node->size, [&](double lng1, double lat1) {
-      e.v0 = S2LatLng::FromDegrees(lat0, lng0).ToPoint();
-      e.v1 = S2LatLng::FromDegrees(lat1, lng1).ToPoint();
-      visit(e);
-    });
-  });
-}
-
-}  // namespace internal
-
-class GeoArrowChain {
- public:
-  explicit GeoArrowChain(const struct GeoArrowGeometryNode* node)
-      : node_(node) {}
-
-  template <typename Visit>
-  void VisitVertices(Visit&& visit) {
-    internal::VisitVertices(node_, visit);
-  }
-
-  template <typename Visit>
-  void VisitEdges(Visit&& visit) {
-    internal::VisitEdges(node_, visit);
-  }
-
- protected:
-  const struct GeoArrowGeometryNode* node_{};
-};
-
-class GeoArrowLoop : public GeoArrowChain {
- public:
-  explicit GeoArrowLoop(const struct GeoArrowGeometryNode* node,
-                        std::vector<S2Point>* scratch)
-      : GeoArrowChain(node), scratch_(scratch) {
-    scratch_->clear();
-  }
-
-  double GetCurvature();
-
- protected:
-  std::vector<S2Point>* scratch_{};
-
-  void BuildScratch();
-};
-
-class GeoArrowGeom {
- public:
-  explicit GeoArrowGeom(struct GeoArrowGeometryView geom) : geom_(geom) {}
-
-  template <typename Visit>
-  void VisitChains(Visit&& visit) {
-    internal::VisitGeoArrowNodes(geom_, [&](struct GeoArrowGeometryNode* node) {
-      visit(GeoArrowChain(node));
-    });
-  }
-
-  template <typename Visit>
-  void VisitLoops(std::vector<S2Point>* scratch, Visit&& visit) {
-    internal::VisitGeoArrowNodes(geom_, [&](struct GeoArrowGeometryNode* node) {
-      visit(GeoArrowLoop(node, scratch));
-    });
-  }
-
-  template <typename Visit>
-  void VisitVertices(Visit&& visit) {
-    VisitChains([&](GeoArrowChain chain) { chain.VisitVertices(visit); });
-  }
-
-  template <typename Visit>
-  void VisitEdges(Visit&& visit) {
-    VisitChains([&](GeoArrowChain chain) { chain.VisitEdges(visit); });
-  }
-
- private:
-  struct GeoArrowGeometryView geom_;
 };
 
 /// @}
